remerging addresses with profile test and adapting to staging env
