+++ tests/e2e/profile-happy-path.test.ts
@@ -9,12 +9,16 @@ import {
   clear,
   link,
   screenshot,
+  waitFor,
+  currentURL,
+  below,
 } from "taiko";
 
 import { authenticateToConnect } from "./utils/authenticate-to-connect";
+import { configVariables } from "@src/configs/config-variables";
 
 describe("Profile Happy path", () => {
-  jest.setTimeout(60000);
+  jest.setTimeout(90000);
 
   beforeAll(async () => {
     await openBrowser({
@@ -34,15 +38,24 @@ describe("Profile Happy path", () => {
     await closeBrowser();
   });
 
-  const isStagingEnv = (process.env.CONNECT_PROFILE_URL || "").includes(
-    "staging",
-  );
+  const isStagingEnv = (
+    process.env.CONNECT_PROFILE_URL || configVariables.connectProfileUrl
+  ).includes("staging");
 
-  test("it should do Profile flows happy path", async (done) => {
-    expect.assertions(isStagingEnv ? 10 : 7);
+  test("it should do Profile and Addresses flows' happy path", async (done) => {
+    expect.assertions(isStagingEnv ? 10 : 20);
 
     try {
       await authenticateToConnect();
+
+      const baseURL = (
+        process.env.CONNECT_TEST_ACCOUNT_URL ||
+        configVariables.connectAccountURL
+      ).replace(/\/?$/, "");
+
+      console.log(isStagingEnv);
+      console.log(process.env.CONNECT_PROFILE_URL);
+      console.log(configVariables.connectProfileUrl);
 
       if (isStagingEnv) {
         expect(await text("Create your profile").exists()).toBeTruthy();
@@ -51,8 +64,8 @@ describe("Profile Happy path", () => {
         expect(await text("Profile | new").exists()).toBeTruthy();
 
         await write(
-          "Superman",
-          into(textBox({ placeholder: "Enter your name" })),
+          "Taiko Test",
+          into(textBox({ placeholder: "Enter your full name" })),
         );
 
         await click("Add my information");
@@ -73,12 +86,123 @@ describe("Profile Happy path", () => {
 
       await clear(textBox("Username"));
       await write(
-        "Superman",
+        "Supertest",
         into(textBox({ placeholder: "Enter your username" })),
       );
 
       await click("Update my information");
       expect(await text("Your profile has been updated").exists()).toBeTruthy();
+
+      // expect(await text("Personal Information").exists()).toBeTruthy();
+      // await click("Personal Information");
+
+      // Add address flow
+      expect(await link("Add new address").exists()).toBeTruthy();
+      await click(link("Add new address"));
+      expect(await text("Address | new").exists()).toBeTruthy();
+
+      await write(
+        "42 Test Road",
+        into(textBox({ placeholder: "Enter your street address" })),
+      );
+
+      await write(
+        "Test City",
+        into(textBox({ placeholder: "Enter your locality" })),
+      );
+      await write(
+        "42420",
+        into(textBox({ placeholder: "Enter your postal code" })),
+      );
+      await write(
+        "Testland",
+        into(textBox({ placeholder: "Enter your country" })),
+      );
+      await write(
+        "Home",
+        into(textBox({ placeholder: "Enter your address kind" })),
+      );
+
+      await click("Add address");
+      await waitFor(350);
+      expect(await currentURL()).toEqual(`${baseURL}/account/profile`);
+
+      if (isStagingEnv) {
+        expect(await link("Add new address").exists()).toBeTruthy();
+        await click(link("Add new address"));
+        expect(await text("Address | new").exists()).toBeTruthy();
+
+        await write(
+          "23 Test Road",
+          into(textBox({ placeholder: "Enter your street address" })),
+        );
+        await write(
+          "Test City",
+          into(textBox({ placeholder: "Enter your locality" })),
+        );
+        await write(
+          "42420",
+          into(textBox({ placeholder: "Enter your postal code" })),
+        );
+        await write(
+          "Testland",
+          into(textBox({ placeholder: "Enter your country" })),
+        );
+        await write(
+          "Work",
+          into(textBox({ placeholder: "Enter your address kind" })),
+        );
+
+        await click("Add address");
+        await waitFor(350);
+        expect(await currentURL()).toEqual(`${baseURL}/account/profile`);
+      }
+
+      // Update address flow
+      await click(link("", below("Addresses")));
+
+      expect(await text("Primary").exists()).toBeTruthy();
+      await click("Update this address");
+      expect(await text("Address | edit").exists()).toBeTruthy();
+
+      await clear(textBox("Street address"));
+      await write(
+        "21b Baker Street",
+        into(textBox({ placeholder: "Enter your street address" })),
+      );
+
+      await click("Update my address");
+      expect(await text("Your address has been updated").exists()).toBeTruthy();
+
+      // Mark address as primary flow
+      await click("Personal Information");
+      expect(await text("Show 1 more").exists()).toBeTruthy();
+      await click("Show 1 more");
+      await click(link("Work"));
+
+      expect(
+        await text("Use this address as my main address").exists(),
+      ).toBeTruthy();
+      await click("Use this address as my main address");
+      expect(
+        await text("You are about to set this address as main.").exists(),
+      ).toBeTruthy();
+      await click("Confirm");
+      expect(
+        await text("Your address has been marked as primary").exists(),
+      ).toBeTruthy();
+
+      // Delete address flow
+      await click(link("", below("Addresses")));
+      expect(await text("Primary").exists()).toBeTruthy();
+
+      await click("Delete this address");
+      expect(
+        await text("You are about to delete this address.").exists(),
+      ).toBeTruthy();
+      await click("Delete");
+      await waitFor(250);
+      expect(await currentURL()).toEqual(`${baseURL}/account/profile`);
 
       done();
     } catch (error) {
